Update jasmine-suites to work with Jasmine 4.6.
jasmine-suites was originally written for Jasmine 1.x, and has managed
to survive up through 3.5. However, it does not work with the latest 4.6
release, due to internal changes with `this` no longer representing the
current suite in a call to `describe()`.

To address this, we simply monkey-patch the `SuiteBuilder`, storing the
current suite being operated on. We can then reference that when needed
during building of the suites.

Along with the fix, the package is now built with Rollup and can be
deployed to npm, with CommonJS and ES module support.

Testing Done:
Tested this previously with Jasmine 4.6 with an attempt to upgrade the
unit test dependencies for Review Board.

Reviewed at https://reviews.reviewboard.org/r/13189/